+++ Utils.py
@@ -19,6 +19,7 @@ import warnings
 
 from argparse import Namespace
 from datetime import datetime, timezone
+from enum import Enum, auto
 
 from settings import Settings, get_settings
 from time import sleep
@@ -1374,3 +1375,12 @@ def get_all_causes(ex: Exception) -> str:
     top = causes[-1]
     others = "".join(f"\n{' ' * (i + 1)}Which caused: {c}" for i, c in enumerate(reversed(causes[:-1])))
     return f"{top}{others}"
+
+
+class Type(Enum):
+    TOOL = auto()
+    MISC = auto()
+    CLIENT = auto()
+    ADJUSTER = auto()
+    FUNC = auto()  # do not use anymore
+    HIDDEN = auto()
